+++ ios/Lift/SensorKind+Localisation.swift
